% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brazil_rapid_transit.R
\name{brazil_rapid_transit}
\alias{brazil_rapid_transit}
\title{Download Brazil's rapid transit corridors and stops}
\usage{
brazil_rapid_transit()
}
\value{
A list with two entries: one sf with the transit stops, and another
one with the transit corridors.
}
\description{
Downloads Brazil's rapid transit corridors and stops made available by ITDP
Brasil at
\href{https://mobilidados.org.br/map-of-rapid-transit-in-brazil}{Mobilidados}.
}
